pyproject.toml
[tool.poetry]
name = "async-web3"
version = "0.1.0"
description = ""
authors = ["Guanqun Lu <guanqunlu@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
websockets = "^8.1.0"
pytest-asyncio = "^0.15.1"
web3 = "^5.19.0"
pyhumps = "^3.0.2"
eth-utils = "^1.10.0"
eth-abi = "^2.1.1"
Brownie = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^5.5"
mypy = "^0.812"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
